move bidara prompts/funcs into src/assistant/
